AB#5033 -- Refactor address validation dialog into reusable components
