+++ frontend/public/locales/fr/common.json
@@ -1,4 +1,22 @@
 {
+  "dialog": {
+    "address-suggestion": {
+      "address-selection-legend": "Veuillez sélectionner une option d'adresse ci\u2011dessous\u00a0:",
+      "cancel-button": "Retour",
+      "description": "Il y a un problème avec l'adresse que vous avez fournie. Veuillez sélectionner l'adresse que vous souhaitez utiliser. Si vous vivez dans un appartement ou une chambre, n'oubliez pas de le préciser.",
+      "entered-address-option": "Adresse saisie\u00a0:",
+      "header": "Suggestion de correction d'adresse",
+      "suggested-address-option": "Adresse corrigée suggérée\u00a0:",
+      "use-selected-address-button": "Utiliser l'adresse sélectionnée"
+    },
+    "address-invalid": {
+      "close-button": "Retour",
+      "description": "Nous n'avons pas pu vérifier votre adresse. Vous pouvez soit la mettre à jour, soit continuer avec l'adresse que vous avez saisie. Si vous vivez dans un appartement ou une chambre, n'oubliez pas d'indiquer cette information.",
+      "entered-address": "Adresse saisie\u00a0:",
+      "header": "Adresse invalide",
+      "use-entered-address-button": "Utiliser l'adresse saisie"
+    }
+  },
   "error-message": {
     "mailing": {
       "address-required": "Entrez l'adresse postale, normalement le numéro et le nom de la rue",
